Create a skeleton web server
This commit contains a basic running web server that serves GET requests correctly and the only missing functionalities are multi threading and POST method.
The reason why POST method wasn't implemented is that it's hard to be tested without the web client because the requirement aren't the same as the standards which prevents testing it using a web browser.